+++ MyCookbook/ConversionMicroserviceAPI/Business/Converters/Volume/ToLiterConverter.cs
@@ -18,7 +18,7 @@ namespace ConversionMicroserviceAPI.Business.Converters
 
         public abstract string UnitFrom { get; }
 
-        public abstract double Convert(double unit);
+        public abstract decimal Convert(decimal unit);
 
         public static void CreateToLiterConverters()
         {
@@ -34,7 +34,7 @@ namespace ConversionMicroserviceAPI.Business.Converters
     {
         public override string UnitFrom => Constants.UNIT_VOLUME_MILLILITER.Name;
 
-        public override double Convert(double milliliters)
+        public override decimal Convert(decimal milliliters)
         {
             return milliliters / 1000;
         }
@@ -44,7 +44,7 @@ namespace ConversionMicroserviceAPI.Business.Converters
     {
         public override string UnitFrom => Constants.UNIT_VOLUME_DECILITER.Name;
 
-        public override double Convert(double deciliters)
+        public override decimal Convert(decimal deciliters)
         {
             return deciliters / 10;
         }
@@ -54,9 +54,9 @@ namespace ConversionMicroserviceAPI.Business.Converters
     {
         public override string UnitFrom => Constants.UNIT_VOLUME_TEASPOON.Name;
 
-        public override double Convert(double teaspoons)
+        public override decimal Convert(decimal teaspoons)
         {
-            return teaspoons * 0.00492892159;
+            return teaspoons * (decimal)0.00492892159;
         }
     }
 
@@ -64,9 +64,9 @@ namespace ConversionMicroserviceAPI.Business.Converters
     {
         public override string UnitFrom => Constants.UNIT_VOLUME_TABLESPOON.Name;
 
-        public override double Convert(double tablespoons)
+        public override decimal Convert(decimal tablespoons)
         {
-            return tablespoons * 0.0147867648;
+            return tablespoons * (decimal)0.0147867648;
         }
     }
 
@@ -74,9 +74,9 @@ namespace ConversionMicroserviceAPI.Business.Converters
     {
         public override string UnitFrom => Constants.UNIT_VOLUME_CUP.Name;
 
-        public override double Convert(double cups)
+        public override decimal Convert(decimal cups)
         {
-            return cups * 0.236588237;
+            return cups * (decimal)0.236588237;
         }
     }
 }
